chore: make domain.py simple to read + rename routes.py to rest_api.py
